backend/requirements.txt
Flask==3.0.3
flask-cors
python-dotenv
PyMySQL
bcrypt
PyJWT
requests
gunicorn
